Fix CUDA/HIP atomics and make lock array functions weak symbols
To allow linking of external desul and internal kokkos desul I made
the lock array functions weak symbols, so that we only get one of them.
I also added missing integral atomics to complement the pure host
GCC atomics available. This was causing compile issues when otherwise
on the host a different overload set exists then on the device.